+++ back/src/main/java/com/bubble/giju/domain/Page/pageController.java
@@ -1,12 +1,12 @@
-package com.bubble.giju.domain.payment.controller;
+package com.bubble.giju.domain.Page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/api/payment")
-public class paymentController2 {
+@RequestMapping("/toss")
+public class pageController {
     @GetMapping("/page")
     public String paymentPage() {
         return "payment";  // payment.html 렌더링
